requirements.txt
langchain>=0.2.1
langchain-core>=0.2.1
langchain-community>=0.2.1
langchain-text-splitters>=0.2.1
langchain-groq>=0.1.5
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=3.0.0
pypdf>=4.0.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
streamlit>=1.35.0
python-dotenv>=1.0.1
tqdm>=4.66.0
openai>=1.30.0